+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - A `railnova_kafka_mtls.py` program to access the Kafka broker using mTLS instead of SASL.
+- A `railnova_kafka_nosr.py` program to access the Kafka broker using mTLS without Schema Registry.
+- All the AVRO schemas required for Railnova's output sharing topics in a `schemas` folder.
 
 ### Changed